+++ lib/src/timeline/timeline_component.dart
@@ -79,14 +79,11 @@ class EssentialTimelineComponent implements OnDestroy, OnInit {
 
   @override
   void ngOnInit() {
-    //print('ngOnInit');
-
     bodyOnClickSubscription = html.document.onClick.listen(bodyOnClick);
   }
 
   @override
   void ngOnDestroy() {
-    //  print('ngOnDestroy');
     bodyOnClickSubscription?.cancel();
     bodyOnClickSubscription = null;
   }
@@ -100,8 +97,7 @@ class EssentialTimelineComponent implements OnDestroy, OnInit {
   }
 
   String getDescription(TimelineRender item, html.HtmlElement element) {
-    element.setInnerHtml(item?.getModel?.description,
-        treeSanitizer: html.NodeTreeSanitizer.trusted);
+    element.setInnerHtml(item?.getModel?.description, treeSanitizer: html.NodeTreeSanitizer.trusted);
     // description = item?.getModel?.description;
     return '';
   }
@@ -111,7 +107,6 @@ class EssentialTimelineComponent implements OnDestroy, OnInit {
   }
 
   String getContentMutedSubtitle(TimelineRender item) {
-    print(item);
     return item.getModel?.contentMutedSubtitle;
   }
 
@@ -130,8 +125,7 @@ class EssentialTimelineComponent implements OnDestroy, OnInit {
   void toggleDropdownOption(e) {
     e.stopPropagation();
     html.HtmlElement element = e.target;
-    var elementSelected =
-        element.closest('.card-footer').querySelector('.dropdown-options');
+    var elementSelected = element.closest('.card-footer').querySelector('.dropdown-options');
 
     if (!elementSelected.classes.contains('show')) {
       elementSelected.classes.add('show');